+++ inc/lights/light_source.hpp
