+++ main.cpp
@@ -1,4 +1,4 @@
-#include "app.hpp"
+#include "app/app.hpp"
 
 int main() {
     Application app;
