+++ src/objects/object.cpp
@@ -2,7 +2,7 @@
 // Created by andrew on 18/04/23.
 //
 
-#include "object.hpp"
+#include "objects/object.hpp"
 
 void Object::setTransformation(const Transformation &transformation) {
     transformation_m = transformation;
