+++ src/objects/plane.cpp
@@ -2,7 +2,7 @@
 // Created by paul on 3/31/23.
 //
 
-#include "plane.hpp"
+#include "objects/plane.hpp"
 
 constexpr double kEpsilon = 0.0001f;
